packages/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devsweep-core"
version = "0.1.0"
description = "Core engine for DevSweep developer storage analysis"
authors = [{name = "DevSweep Team"}]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.12.0",
    "psutil>=5.9.0",
    "litellm>=1.0.0",
    "dynaconf>=3.2.0",
    "structlog>=23.1.0",
    "anyio>=4.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.80.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.5.0",
    "pyinstaller>=6.10.0",
    "ruff>=0.1.0",
    "taskipy>=1.12.0",
]

[project.scripts]
devsweep-sidecar = "devsweep.api.__main__:main"

[tool.taskipy.tasks]
format = "ruff format ."
lint = "ruff check . --fix"
typecheck = "mypy ."
test = "pytest"
test-unit = "pytest tests/unit"
test-integration = "pytest tests/integration"
test-property = "pytest tests/property"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "PT"]
ignore = ["ANN101", "ANN102"]

[tool.hatch.build.targets.wheel]
packages = ["devsweep"]

[tool.mypy]
strict = true
python_version = "3.12"
